+++ MTP/application/services/MTPNetworkController.swift
@@ -92,6 +92,8 @@ enum MTP: Hashable {
     case userGetByToken
     /// userPosts(id: Int)
     case userPosts(id: Int)
+    /// userPost(payload: UserUpdatePayload)
+    case userPost(id: Int, token: String)
     /// userPut(payload: UserUpdatePayload)
     case userPut(payload: UserUpdatePayload)
     /// userLogin(email: String, password: String)
@@ -176,12 +178,14 @@ extension MTP: TargetType {
         case .userDelete(let id),
              .userGet(let id):
             return "user/\(id)"
-        case .userPut(let payload):
-            return "user/\(payload.id)"
+        case .userPost(let id, _):
+            return "user/\(id)/token"
         case .userPosts(let id):
             return "users/\(id)/location-posts"
         case .userLogin:
             return "user/login"
+        case .userPut(let payload):
+            return "user/\(payload.id)"
         case .userRegister:
             return "user"
         case .userVerify(let id):
@@ -226,6 +230,7 @@ extension MTP: TargetType {
              .postPublish,
              .upload,
              .userLogin,
+             .userPost,
              .userRegister:
             return .post
         case .userPut:
@@ -267,6 +272,10 @@ extension MTP: TargetType {
         case let .search(query?):
             return .requestParameters(parameters: ["query": query],
                                       encoding: URLEncoding.default)
+        case .userPost(_, let token):
+            return .requestParameters(parameters: ["type": "apn_device_token",
+                                                   "value": token],
+                                      encoding: URLEncoding(destination: .queryString))
         case let .upload(photo: photo, caption: caption, location: location):
             let filePart = MultipartFormData(provider: .data(photo),
                                              name: "files[0]",
@@ -366,6 +375,12 @@ extension MTP: TargetType {
             file = "\(self)"
         //case let .photos(user?, page):
             //file = "photos-\(user)-\(page)"
+        case .locationPhotos:
+            //file = "locationPhotos-\(location)"
+            file = "locationPhotos-554"
+        case .locationPosts:
+            //file = "locationPosts-\(location)"
+            file = "locationPosts-554"
         case .photos(_, let page):
             file = "photos-7853-\(page)"
         case .rankings:
@@ -376,6 +391,8 @@ extension MTP: TargetType {
         case .userGet:
             //file = "userGet-\(id)"
             file = "userGet-1"
+        case .userLogin:
+            file = "userLogin-7853"
         case .userPosts:
             //file = "userPosts-\(id)"
             file = "userPosts-7853"
@@ -419,6 +436,7 @@ extension MTP: AccessTokenAuthorizable {
              .upload,
              .userDelete,
              .userGetByToken,
+             .userPost,
              .userPut,
              .userVerify:
             return .bearer
@@ -591,6 +609,10 @@ struct MTPNetworkController: ServiceProvider {
     }
 
     /// Load checklists
+    ///
+    /// - Parameters:
+    ///   - stub: Stub behaviour
+    ///   - then: Completion
     func loadChecklists(stub: @escaping MTPProvider.StubClosure = MTPProvider.neverStub,
                         then: @escaping NetworkCompletion<Checked> = { _ in }) {
         guard data.isLoggedIn else {
@@ -757,11 +779,13 @@ struct MTPNetworkController: ServiceProvider {
     /// - Parameters:
     ///   - id: Location ID
     ///   - reload: Force reload
+    ///   - stub: Stub behaviour
     ///   - then: Completion
     func loadPhotos(location id: Int,
                     reload: Bool,
+                    stub: @escaping MTPProvider.StubClosure = MTPProvider.neverStub,
                     then: @escaping NetworkCompletion<PhotosInfoJSON> = { _ in }) {
-        let provider = MTPProvider()
+        let provider = MTPProvider(stubClosure: stub)
         let endpoint = MTP.locationPhotos(location: id)
         guard reload || !endpoint.isThrottled else {
             return then(.failure(.throttle))
@@ -803,6 +827,7 @@ struct MTPNetworkController: ServiceProvider {
     /// - Parameters:
     ///   - page: Index
     ///   - reload: Force reload
+    ///   - stub: Stub behaviour
     ///   - then: Completion
     func loadPhotos(page: Int,
                     reload: Bool,
@@ -821,6 +846,7 @@ struct MTPNetworkController: ServiceProvider {
     ///   - id: User ID
     ///   - page: Index
     ///   - reload: Force reload
+    ///   - stub: Stub behaviour
     ///   - then: Completion
     func loadPhotos(profile id: Int,
                     page: Int,
@@ -889,10 +915,12 @@ struct MTPNetworkController: ServiceProvider {
     ///
     /// - Parameters:
     ///   - id: Location ID
+    ///   - stub: Stub behaviour
     ///   - then: Completion
     func loadPosts(location id: Int,
+                   stub: @escaping MTPProvider.StubClosure = MTPProvider.neverStub,
                    then: @escaping NetworkCompletion<PostsJSON> = { _ in }) {
-        let provider = MTPProvider()
+        let provider = MTPProvider(stubClosure: stub)
         let endpoint = MTP.locationPosts(location: id)
         guard !endpoint.isThrottled else {
             return then(.failure(.throttle))
@@ -933,6 +961,7 @@ struct MTPNetworkController: ServiceProvider {
     ///
     /// - Parameters:
     ///   - id: User ID
+    ///   - stub: Stub behaviour
     ///   - then: Completion
     func loadPosts(user id: Int,
                    stub: @escaping MTPProvider.StubClosure = MTPProvider.neverStub,
@@ -983,6 +1012,7 @@ struct MTPNetworkController: ServiceProvider {
     ///
     /// - Parameters:
     ///   - query: Filter
+    ///   - stub: Stub behaviour
     ///   - then: Completion
     func loadRankings(query: RankingsQuery,
                       stub: @escaping MTPProvider.StubClosure = MTPProvider.neverStub,
@@ -1070,6 +1100,7 @@ struct MTPNetworkController: ServiceProvider {
     /// - Parameters:
     ///   - list: Checklist
     ///   - id: User ID
+    ///   - stub: Stub behaviour
     ///   - then: Completion
     func loadScorecard(list: Checklist,
                        user id: Int,
@@ -1190,6 +1221,7 @@ struct MTPNetworkController: ServiceProvider {
     ///
     /// - Parameters:
     ///   - id: User ID
+    ///   - stub: Stub behaviour
     ///   - then: Completion
     func loadUser(id: Int,
                   stub: @escaping MTPProvider.StubClosure = MTPProvider.neverStub,
@@ -1586,15 +1618,17 @@ struct MTPNetworkController: ServiceProvider {
     /// - Parameters:
     ///   - email: Email
     ///   - password: Password
+    ///   - stub: Stub behaviour
     ///   - then: Completion
     func userLogin(email: String,
                    password: String,
+                   stub: @escaping MTPProvider.StubClosure = MTPProvider.neverStub,
                    then: @escaping NetworkCompletion<UserJSON>) {
         guard !email.isEmpty && !password.isEmpty else {
             return then(.failure(.parameter))
         }
 
-        let provider = MTPProvider()
+        let provider = MTPProvider(stubClosure: stub)
         let endpoint = MTP.userLogin(email: email, password: password)
 
         func parse(success response: Response) {
@@ -1703,6 +1737,46 @@ struct MTPNetworkController: ServiceProvider {
                     } else {
                         return then(.failure(.message(reply.message)))
                     }
+                } catch {
+                    self.log.error("decoding: \(endpoint.path): \(error)\n-\n\(result.toString)")
+                    return then(.failure(.decoding))
+                }
+            case let .failure(.underlying(error, response)):
+                let problem = self.parse(error: error, response: response)
+                return then(.failure(problem))
+            case .failure(let error):
+                guard error.modified(from: endpoint) else {
+                    return then(.failure(.notModified))
+                }
+                let message = error.errorDescription ?? L.unknown()
+                self.log.error("failure: \(endpoint.path) \(message)")
+                return then(.failure(.network(message)))
+            }
+        }
+    }
+
+    /// Update user token
+    ///
+    /// - Parameters:
+    ///   - token: String
+    ///   - then: Completion
+    func userUpdate(token: String,
+                    then: @escaping NetworkCompletion<UserTokenReply>) {
+        guard let userId = data.user?.id else {
+            return then(.failure(.parameter))
+        }
+
+        let auth = AccessTokenPlugin { self.data.token }
+        let provider = MTPProvider(plugins: [auth])
+        let endpoint = MTP.userPost(id: userId, token: token)
+
+        provider.request(endpoint) { response in
+            switch response {
+            case .success(let result):
+                do {
+                    let reply = try result.map(UserTokenReply.self,
+                                               using: JSONDecoder.mtp)
+                    return then(.success(reply))
                 } catch {
                     self.log.error("decoding: \(endpoint.path): \(error)\n-\n\(result.toString)")
                     return then(.failure(.decoding))